requirements.txt
asttokens==3.0.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.1.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
matplotlib-inline==0.1.7
mysql-connector-python==9.3.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
regex==2024.11.6
six==1.17.0
SQLAlchemy==2.0.40
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
